Assignment 1
Generate a 10x10 grid of Unity Cubes. Each cube should be a Gameobject and attached to it
should be a script which has information about that particular tile.
After that, do a ray cast from the mouse. When the mouse hovers over the tile, you should read
that tile’s information using the ray cast and display the unit’s position on the grid on a UI
element.