Assignment 3
Generate a player unit on the map. The player should be able to move to any selected tile on the
grid generated in Assignment 1 and should not be able to move to grids with Obstacles if
Assignment 2 was attempted. Please don’t use Unity Pathfinding for this. A grid-based
pathfinding algorithm is expected. Please show movement when moving from one tile to
another. Input should be disabled while the unit is moving.